refactor: move dtype config out of base
