+++ src/unilab/envs/locomotion/common/commands.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, field
 
 import numpy as np
 
-from unilab.base.dtype_config import get_global_dtype
+from unilab.dtype_config import get_global_dtype
 
 
 @dataclass
