+++ src/unilab/envs/manipulation/inhand_rot_allegro/base.py
@@ -7,8 +7,8 @@ import numpy as np
 
 from unilab.base.backend import SimBackend
 from unilab.base.base import EnvCfg
-from unilab.base.dtype_config import get_global_dtype
 from unilab.base.np_env import NpEnv, NpEnvState
+from unilab.dtype_config import get_global_dtype
 
 
 @dataclass
