+++ src/unilab/envs/manipulation/sharpa_inhand/base.py
@@ -10,8 +10,8 @@ import numpy as np
 from unilab.assets import ASSETS_ROOT_PATH
 from unilab.base.backend import SimBackend
 from unilab.base.base import EnvCfg
-from unilab.base.dtype_config import get_global_dtype
 from unilab.base.np_env import NpEnv, NpEnvState
+from unilab.dtype_config import get_global_dtype
 from unilab.envs.common.rotation import np_quat_apply, np_quat_mul
 
 DEFAULT_ACTUATED_JOINT_NAMES: list[str] = [
